Almaz/requirements.txt
aiogram==2.25.2
aiohttp==3.8.6
python-dotenv==1.0.1